Implement a pdo-shell that supports contract-specific plugins (#15)
* Implement a pdo-shell that supports contract-specific plugins

This commit adds the pdo-shell interpreter for interacting with
private data objects. The shell provides some basic commands for
configuring provisioning and enclave services, managing identities,
creating & sending messages to contract objects.

In addition, the shell can be extended to provide per-contract class
extensions through a plugin system. And example plugin for the
integer-key contract is provided.

Finally, the shell can be scripted for execution through the command
line or through the shell prompt itself.

The details of the shell are still in flux. Documentation will follow
when the interface stabilizes.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix import for integer-key plugin

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix typo in ps-start.sh

Signed-off-by: Byron Marohn <byron.marohn@intel.com>