Add a way for contracts to specify methods that do not change state.
For methods that do not change state, we can completely avoid state
serialization. For now, the state that is returned is "empty" except for
the initialization vectors. Clients need to identify that and not submit
to the ledger. Correctly implementing this will require the TP to
identify empty state updates and reject them.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>