Clean up some of the enclave debugging to make it less verbose
Dropped the logging level to debug for all of the statements in the
secrets module. Add the "debug" prefix to any variables that are created
exclusively for logging.

Added a critical message in enclave initialization to note that the
enclave was compiled with debugging on. Debug logging creates an obvious
channel that exposes confidential information.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>