Canonicalize handling of the LEDGER_URL environment variable for configuration.
Fix a format problem with LEDGER_URL in the auction test configuration.
Add support in all of our scripts for expansion of ${ledger} in the
configuration files. This will be set by default to http://127.0.0.1:8008
unless the environment variable $LEDGER_URL is set (which is part of our
documentation). This should simplify start up for all of the scripts.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>