Implements a number of performance related fixes (#47)
* Add a way for contracts to specify methods that do not change state.

For methods that do not change state, we can completely avoid state
serialization. For now, the state that is returned is "empty" except for
the initialization vectors. Clients need to identify that and not submit
to the ledger. Correctly implementing this will require the TP to
identify empty state updates and reject them.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add support to contract response for operations that do not modify state.

Check the result of contract execution for an empty state that signfies
that no change was made. Short circuit the response to avoid encryption
and signing.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Optimize construction of the gipsy intiailization packages.

Adds a script that calls tinyscheme to read all of the package files
into a single file, removing all whitespace and comments.  Currently
this reduces the size of the initial package by about 40%.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Adjust the initial memory allocation for the interpreter to reduce garbage collection.

Bump up the amount of memory initially allocated to the interpreter. This reduces
the amount of garbage collection that is required for many contracts.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Optimize serialization of state in the interpreter.

Rewrite a few of the serialization functions for the gipsy
interpreter. Pretty substantial performance benefits for
vectors and lists.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Update the test functions so that state is only updated if it is changed.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Push learnings about writing efficient contracts into the code.

This commit 1) sets many of the contract methods to be constants so that
state will not be saved by the enclave, 2) fixes the clients so that
they don't try to push state that doesn't exist, and 3) changes how
contracts are initialized so that we don't duplicate initialization
every time we restore state in the enclave.

Result is that in-enclave time has been reduced for integer-key tests by
more than 60%.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add a timer call that can be used for performance testing.

Note this is a very simple facility and very coarse grained.

As with other enclave utilities timer code is compiled into the enclave
only when debug is turned on. A call to "__TIMEIT__()" (which is a
macro) will generate timing information for the function in which it was
invoked.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Make the GetTimer function available outside the enclave.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Update the eservice test contracts with the variable initialization and const method changes.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Remove timing calls.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>