Implemented enclave block store
The block store is a state cache for the eservice which eliminates the
need to send state back and forth on every transaction. State is pulled
from the block store when running contracts, and the resulting state is
placed there when the contract execution completes.

Current implementation details:
- The block store is not persistent across restarts
- The client must retrieve state from the block store to store on the
ledger after any operation that updates the state
- The client must put the state block(s) into the block store when
starting work with an enclave service that does not currently have them.

Signed-off-by: Byron Marohn <byron.marohn@intel.com>