Update the gipsy kv operations to use the unprivileged calls
Moved the interpreter_kv code out of the enclave and into
common so we can use it in the untrusted part of gipsy.

Made a number of changes to manage memory a little better. Dropped
the state hash concept and adopted bruno's block id concept. This
cleaned up a lot of the oddness in parameters.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>