remove the redundant state arguments for the interpreter... the KV borg has assimilated us! (#97)
* reformat the interpreter_kv file and clean up some of the function dfns

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix the double definition of SAFE_LOG and make it DEBUG dependent for all

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Only the KV store is passed into the interpreter removing the redunandant args

LOTS of clean up in here. Some necessary to remove all the old redundant (and
in some cases broken) uses of the string state.

Format cleanup in several places.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add gipsy interpreter functions to access the key value store directly

Adds functions to get, set, and delete key/value pairs from the
persistent key value store. In untrusted mode, add functions to
open and close a key value store.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* add tests for the persistent key/value store contract ops

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* update the integer key and auction contracts to use the key value store

This commit includes a wrapper for the persistent key/value store
gipsy operations to track keys that are stored. Integer-Key and
Auction contracts have been updated to use this. So the keys remain
in the intrinsic state but the values are all pushed into the kv
store.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Update the gipsy kv operations to use the unprivileged calls

Moved the interpreter_kv code out of the enclave and into
common so we can use it in the untrusted part of gipsy.

Made a number of changes to manage memory a little better. Dropped
the state hash concept and adopted bruno's block id concept. This
cleaned up a lot of the oddness in parameters.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix the unprivileged delete call for the interpreter kv store

Signed-off-by: Mic Bowman <mic.bowman@intel.com>