Update configuration and directory structures to move all  file handling to a single location
Move the eservice test contracts into the contracts directory tree

Update the handling of data_dir to ensure the configured directory is used.
There is a new function in pdo.common.utility called set_default_data_directory
that should be called once the data directory configuration is set. This will
be used by the state cache, contract cache and others.

Update the client scripts to use the data directory

Update the service scripts to handle the data directory

Signed-off-by: Mic Bowman <mic.bowman@intel.com>