Make logging more consistent
Replaced several of the old "Log" statements with SAFE_LOG
statements. Moved the untrusted logger out of the state tests
and into the root of the common directory (though it will only
be built when building untrusted code). Cleaned up some old
includes in the makefiles. And made use of DEBUG a little more
consistent.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>