Several updates that fix a couple bugs and make contracts more resilient (#129)
* Check for errors in the return value for key-value-store operations.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Make logging more consistent

Replaced several of the old "Log" statements with SAFE_LOG
statements. Moved the untrusted logger out of the state tests
and into the root of the common directory (though it will only
be built when building untrusted code). Cleaned up some old
includes in the makefiles. And made use of DEBUG a little more
consistent.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix SchemeExtensions to work with the modified key value store constructors

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Make exchange contract tests mandatory for installation

The tests in the exchange contract family are now run as part
of the installation package. In particular, this should ensure
that the pcontract.so is built correctly. It will also test
some of the untrusted state operations.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Reimplement the common logging module and push throughout the project.

Moved the eservice/pservice configurable logging code into
common. Default logger prints to standard out. Enclave SWIG
wrappers set the logging function to push messages to the
python logger.

This really means that the common utilities can safely use logging
without the fear of having multiple symbols defined (or in some
cases no logging symbols defined).

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Replace 5 parameter vsnprintf with the 4 parameter version.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Update the contract makefiles to recognize test target

Signed-off-by: Mic Bowman <mic.bowman@intel.com>