Implement size limits for memory used and intrinsic state.
Add support for a hard limit on the total amount of memory used
by the gipsy interpreter. Also limit the maximum size of the
intrinsic state.

Bump up the amount of heap memory allocated to the enclave to 16M.

Split constructor/destructor operations out into a Initialize
and Finalize procedures that can be called by the worker thread
to clean up the interpreter. Invoke these in the contract
worker thread.

Update the memory test with expected results and include it
in the run-tests.sh script. Update the test-contract.py
script to handle failure results better and also to handle
very large responses correctly.

Fix a macro define in the log.h

Signed-off-by: Mic Bowman <mic.bowman@intel.com>