Implement size limits for interpreter and intrinsic state (#137)
* Implement size limits for memory used and intrinsic state.

Add support for a hard limit on the total amount of memory used
by the gipsy interpreter. Also limit the maximum size of the
intrinsic state.

Bump up the amount of heap memory allocated to the enclave to 16M.

Split constructor/destructor operations out into a Initialize
and Finalize procedures that can be called by the worker thread
to clean up the interpreter. Invoke these in the contract
worker thread.

Update the memory test with expected results and include it
in the run-tests.sh script. Update the test-contract.py
script to handle failure results better and also to handle
very large responses correctly.

Fix a macro define in the log.h

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add the helper shutdown to test-secrets so the enclave exits cleanly

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Implement realloc for scheme and push use into the interpreter.

Implement a safe_realloc_for_scheme and use it in the tinyscheme
interpreter. Fixed a couple places where memory allocation was inefficient
in tinyscheme.

Replace all StringArray's with ByteArray's in the GipsyInterpreter. Should
remove some very big copies.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Remove memory tests that cannot succeed with out a block cache.

The block cache has not been implemented at this point. Several of the
key value store memory tests fail because we run out of space in the
"not cache".

Also adds comments to the memory test expressions file.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>