Refactor tests and move test scripts into the build directory (#138)
* Refactor tests and move test scripts into the build directory

Several changes... first, moved the test-request and test-contract
scripts out of eservice and into the python directory tree in
pdo/test. These are then installed as pdo-test-request and
pdo-test-contract.

Moved all of the scripts for test-contract out of the eservice tree and
put them into a directory in the build tree.

Added several new tests and restructured the existing ones. Two main
changes. First, added invocation of unit tests for the main packages
that support them (common, python, eservice, and contract). The
enclave-only test-secrets and test-requests are invoked as unit tests
for eservice. Second, changed the enclave configuration so that each
eservice used in the test will use its own data file. Added a "random"
enclave switch to the client update script to randomly pick an
enclave. All this makes it possible to test multiple independent
eservices and to ensure that the client is moving state blocks between
them correctly.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Drop automatic execution of common tests during the build.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Remove flags that prevent tests from running. Earlier solution addresses this more cleanly.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>