Add some classes to map common data structures to the KV store.
Add gipsy packages for a persistent vector, a persistent queue, and
a persistent set. These packages use minimal intrinsic state to store
the bulk of the data in the KV store. Supports most of the standard
Scheme iterators (map, for-each, foldr).

Add a generalized method for converting strings to "stuff".

Signed-off-by: Mic Bowman <mic.bowman@intel.com>