Various updates to improve usefulness and robustness of contracts (#146)
* Enable serialization of closures in contract objects.

Also ensure that procedures are not serialized. Any procedure can
be serialized simply by wrapping it in a closure.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Clean up test and debug contracts correctly

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Split the contract base, signing, encryption and utility classes into separate files.

Partly in the spirit of consistency with one package per file but also to
avoid the need to include lots of unnecessary stuff, moved all of the
functions that had previously been in the contract base file into
separate files.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix the error reporting in the debug utilities.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Update calls to coerce number to use the new utility package

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add some classes to map common data structures to the KV store.

Add gipsy packages for a persistent vector, a persistent queue, and
a persistent set. These packages use minimal intrinsic state to store
the bulk of the data in the KV store. Supports most of the standard
Scheme iterators (map, for-each, foldr).

Add a generalized method for converting strings to "stuff".

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix memory leak for string ports in tinyscheme.

Update memory test to with a vector-based test for
adding and removing elements.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Update the memory test to use the vectors to test adding and removing slots

Signed-off-by: Mic Bowman <mic.bowman@intel.com>