Implement the storage service (#148)
This implements a storage service designed to run side-by-side with the
enclave service. The storage service shares an LMDB database with the
enclave service.

Several differences between the storage service and the storage interface
implemented in the enclave service. First, the storage service move binary
data for blocks in order to avoid the multiple levels of format translation
that occurs with the existing interfaces. Second, the storage service
provides bulk upload and status checks consolidating a number of operations
into a single call. Download is still one block at a time. Finally, the
storage service provides a signature for a bulk upload that can be used for
proof of replication.

This update does not integrate calls to the storage service into the clients
though it does provide a script to test the storage service API.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>