Add support for python expression expansion in configuration templates.
This adds (and then builds on) expressions in the configuration
files. Any string of the form '${{expr}}' will result in the
text inside the braces being evaluated using the current configuration
variable map.

Removed the explicit "http-base" from the configuration tool in favor
of expressions and the '_count_' variable which is more universal.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>