Random fixes.
Several places where i didn't update encrypted_state to raw_state. That
is now fixed.

Fixed the state reordering test for the binary state representation.
Note that negative tests need to be checked carefully to make sure
that they fail for the right reasons.

Moved the state reording into the helpers and updated the MANIFEST
to include that file.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>