First attempt to make the EService more reliable during concurrent transactions.
While this doesn't fix the problem that the client appears to drop the
connection to the server before retrieving the response, it makes all
of the behaviors of the EService twisted framework exposable through the
log messages. This also upgrades some of the function calls used for
errors, connection requests, and protocols.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>