Convert EService to WSGI framework.
Primary change: refactor the handlers in the
EService script to use the python standard WSGI
format. Use the twisted WSGI handlers for dispatch
of requests. This did require that we move back to
base64 encoding for the invocation requests pending
resolution of an issue with binary encoding in the
python cgi parsing library. There are now two entry
points in the enclave swig for processing invocations,
one that is binary and one that is base64 encoded.

Added a MANIFEST to the eservice similar to how
we set up the python files in the python tree.
Makes managing dependencies a little easier.

Added a session and request identifier into every
message to the enclave in order to simplify debugging
of the requests.

Remove a bug with twisted logs.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>