A variety of performance and stability fixes (#155)
* Contract invocation is binary end-to-end.

Removes the json and base64 encoding used to move
the encrypted request and response.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Store the contract code in the contract state so that it only needs to be sent once.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* First attempt to make the EService more reliable during concurrent transactions.

While this doesn't fix the problem that the client appears to drop the
connection to the server before retrieving the response, it makes all
of the behaviors of the EService twisted framework exposable through the
log messages. This also upgrades some of the function calls used for
errors, connection requests, and protocols.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add the perf test script

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Convert EService to WSGI framework.

Primary change: refactor the handlers in the
EService script to use the python standard WSGI
format. Use the twisted WSGI handlers for dispatch
of requests. This did require that we move back to
base64 encoding for the invocation requests pending
resolution of an issue with binary encoding in the
python cgi parsing library. There are now two entry
points in the enclave swig for processing invocations,
one that is binary and one that is base64 encoded.

Added a MANIFEST to the eservice similar to how
we set up the python files in the python tree.
Makes managing dependencies a little easier.

Added a session and request identifier into every
message to the enclave in order to simplify debugging
of the requests.

Remove a bug with twisted logs.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* fix map size: 1GB --> 1TB

Signed-off-by: Bruno Vavala <bruno.vavala@intel.com>