Refactor storage service using WSGI applications (#157)
* Clean up old code left in the eservice

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Refactor storage service for WSGI interface

Refactor the storage service in the same way that we
did earlier for the enlave service. Move all functions
into WSGI applications.

Pulled the WSGI common modules out of the eservice
so we could reuse them in the storage service. Updated
the eservice WSGI applications to use the new module.

Updated the parsing of multipart/form data to use the
encoders and decoders from requests_toolbelt. This greatly
simplifies the code (from the stock python cgi module) and
appears to make it possible to once again send binary
invocation data.

Which... made it possible to add the multi-block get
application for the storage manager. Should reduce the
number of operations on the storage manager.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>