Break out the build for keys and configuration files (#200)
* Fix dependency for pcontract.toml

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Break out the build for keys and configuration files

Currently, every time we build we overwrite the keys and configuration
files in the install directory tree.  This is rather inconvenient if you
have contracts that you want to keep running. This only builds the keys
and configuration files if they are missing or if the templates have
changed.

I'm sure there is other Makefile wizardry that could build implicit
rules for configuration files, but this works.

Signed-off-by: Mic Bowman <cmickeyb@gmail.com>