Major overhaul of the pdo-shell
The shell is now a shell.

Process errors consistently and correctly. All commands should raise an
appropriate exception on failure. If in an interactive shell this will
allow future commands. If in a non-interactive shell, it will be
terminated with an exit condition set that can be tests.

Add if/else/fi conditional construct. This isn't very syntactically
pretty, but it does make it possible to test and handle error conditions
more gracefully.

Variable expansion happens after partitioning into arguments. This makes
it much easier to handle variables that expand to strings with spaces.

Add explicit argument types for scheme_string and scheme_expr to
simplify passing the right structure. This simplifies a lot of the
multiple quoting problems.

Restructure the exchange contracts.
Restructure the integer key contracts.
Restructure the integer key auction contracts.

Fix the enclave parameter for the send command

Signed-off-by: Mic Bowman <mic.bowman@intel.com>