Remove the add, update and auction test scripts. Replace with PSH scripts.
Replace the pdo-create, pdo-update and pdo-auctiontest scripts with
corresponding pdo-shell scripts. Update run-tests.sh with the new
scripts.

Had to make some changes to the way the shell handles command line arguments
for creating variable bindings. Scripts can now use "--key value" to
bind the value to a key. Will remove the "-m" form in a later commit.

Remove all vestiges of "one contract" from the client configuration file to
generalize the configuration for "all" contracts the client might run.

Updated the handling of the contract configuration in the test
procedures to set appropriate defaults.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>