Overhaul the eservice database to simplify the code and the API
Fairly major overhaul of the eservice database code. Removes a lot
of the complexity caused by some extra API functions that were not
likely to be used. Old functions are supported by marked as deprecated
(meaning that they will generate a warning if used).

Pushed the updated API into the applications that use it.

And fixed several problems with makefiles, manifests and such.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>