Include the Wawaka interpreter.
To build for Wawaka you must set the environment variable PDO_CONTRACT_INTERPRETER
to Wawaka. To build for Gipsy, set it to (surprisingly) Gipsy. One interpreter per
compiled enclave. This is not the final solution but will suffice for the moment.

Add the remaining crypto functions to the WASM native interface

Add the RSA and AES functions to the WASM extensions. Also changed
the way the functions are declared (greatly simplifying the code).

Integrate the wawaka interpreter into the build process to simplify setup

Set the environment variable PDO_INTERPRETER to either gipsy (the default)
or wawaka to enable and build a particular interpreter.

Update for the next api to retrieve interpreter opaque data.

add strstr to wawaka extensions

Signed-off-by: Mic Bowman <mic.bowman@intel.com>