Address PR feedback
Add to the code documentation for the s-expression format
that can converted to JSON.

Reverse the order of processing object fields to preserve
the ordering in the expression. In theory JSON is orderless.
In practice its a lot easier to write tests if the order is
preserved.

Add a bunch of additional complex tests and tests for validation.

Fix documentation for path to gipsyscheme interpreter for tests

Signed-off-by: Mic Bowman <mic.bowman@intel.com>