Implement the invocation request interface in Wawaka
Push support for the formal JSON RPC specification into the Wawaka
interpreter. This requires a few changes to the interpreter and a large
number of changes to the dispatch supporting routines.

One thing still missing and out of scope for this update: wawaka
contracts do not have a way to access positional parameters.

This commit also completes addition of dependency support for Wawaka
contracts.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>