Update wawaka support for WAMR-03-05-2020
Update the wawaka to support WAMR tag release WAMR-03-05-2020. Several
changes to the build process and to the integration of native functions.
There are still opportunities to take advantage of the parameter passing
for pointers, however I think it best to keep the validation of pointers
that we have.

Must get the correct submodule for WAMR.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>