Update a number of the wawaka contract classes to improve memory management
Add wrappers for the crytography and encoding extension functions
to use StringArray correctly.

Isolate all handling of memory from extension functions in a copy
from internal pointers. Currently this is set for safe mode (meaning
that we copy and free immediately) though a performant mode that
moves management of the pointers into StringArray.

Improve memory management in StringArray and Value.

Add "dot" operations for object retrieval in Value.
Add structures to Value.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>