Make the use of PDO_LEDGER_KEY_ROOT consistent
PDO_LEGER_KEY_ROOT is used in different ways for
CCF and Sawtooth. But there is no reason why they
need to be in different locations. This unifies
the use of PDO_LEDGER_KEY_ROOT. If you want the CCF
keys in the CCF directory tree then change the
variable when you build and install the PDO TP.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>