Replace use of HOSTNAME with PDO_HOSTNAME
HOSTNAME is a fairly standard environment variable and
we should not be suggesting changes to its value. This
converts all references to HOSTNAME with PDO_HOSTNAME
so we can use it for specifying the interface to use
for services.

Fixes a couple configuration files and sets the
default value in common-config.

Significant update to the CCF PDO TP readme to
reference the existing environment variables and
to make markdown formatting more consistent.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>