Add the exchange contract family for wawaka.
This commit adds the first set of contracts
for the exchange family of contracts. While there are
some changes, the documentation for the exchange contract
family for gypsy is still fundamentally correct. The specific
apis, however, have changed and are documented in the doc
directory.

A test script is available in the scripts directory. Run
"run-tests.sh" after you have installed the contracts.

Although the issuer method claim and disburse are fully
implemented, they have not be tested. Testing will wait
until the exchange and auction contracts are complete.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>