Update the shell to handle errors and subprocesses more cleanly (#315)
Add the ability to handle errors in subcommands. The command "trap_error"
will convert errors in commands from exceptions to successful returns. If
an error does occur, then the symbols "_error_code_" and "_error_message_"
will be set with the numeric code (generally -1) and error message from
the failed command. This is critical for writing negative tests (that is,
writing tests that are supposed to fail).

Enable the creation of script local bindings. In the past, the "script"
command operated on the current bindings. as a result, any changes in
the bindings made in the child script were transparently propogated to
the parent script. The new shell generally preserves that behavior, but
introduces a script local format for variables. In this case, any variable
that starts with '_' is considered a script local variable. This means that
any "global" variable may be changed at random by a sub-script. But that
script local variables will not change.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>