Make elliptic curve parameterizable
This allows to define the elliptic curve used by the PDO crypto via the
-DPDO_USE_ECDSA_CURVE flag during compile time. When not defined, the
default curve is secp256k1 as used in BTC. This commit is motivated by
the use of the PDO crypto with Fabric Private Chaincode. In Fabric,
however, the default ellipctic curve is secp256r1 (aka prime256v1). This
compile flag allows to change the curve when used with FPC to be
complient with Fabric.

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>