Separate update and initialize into separate APIs in the eservice [skip ci]
Separate the code for the initialize and update operations in the
eservice enclave and eservice wrapper code. This introduces a new
request type for initialize, a new response type for initialize and a
new response type for an update that does not modify state. In the case
of the latter, we were already returning a subset of the response fields
so this just makes explicit what we were already doing.

Enable access to the APIs through the actual eservice code.

NOTE: this will not run!

Signed-off-by: Mic Bowman <mic.bowman@intel.com>