Implements contract metadata and registers it with CCF
When a contract is initialized an encryption key pair and a signing key
pair are created and stored in privileged keys in state. A hash is
computed and passed through the contract initialization to CCF. CCF can
then provide an attestation to the veracity of the metadata hash for a
contract.  This ultimately enables contract to contract attestation.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>