Implment update contract & ledger commands for pdosh
Implement commands for interacting with the ledger (e.g.
to get the ledger verification key or ledger attested information
about the contracts or enclaves) and with a contract (e.g.
to get the contract id or provisioned enclaves). This enables
scripts to collect the information necessary to create a verifiable
attestation for a contract.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>