Enable contract to contract attestation (#329)
* Update wawaka crypto and key/value routines

Add crypto_hash to the exported function in the Wawaka Cryptography
suite of functions.

Add privileged get methods for the key value store. These operations
are read-only; privileged keys may be read by the contract but may
not be set.

Remove extra null terminators from the keys. Since these are managed
as std::string's the c_str() accessor will add the null terminator
when necessary.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Improve handling and caching of connections to the ledger

Improve the handling of ledger caches so that we can have
connections to more than one ledger if necessary.

Add a submitter operation to get information about the ledger
if it is available. For CCF this translates into getting the
ledger verifying key.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Split and rename verifying_contract CCF API

Split verify_contract_info into two operations, one for contract
metadata and one for contract provisioning information. The purpose is
to keep the relatively small contract metadata data separate from the
relatively complex provisioning information (which contains all of the
enclave information). This will simplify verification of the contract
metadata.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Separate update and initialize into separate APIs in the eservice [skip ci]

Separate the code for the initialize and update operations in the
eservice enclave and eservice wrapper code. This introduces a new
request type for initialize, a new response type for initialize and a
new response type for an update that does not modify state. In the case
of the latter, we were already returning a subset of the response fields
so this just makes explicit what we were already doing.

Enable access to the APIs through the actual eservice code.

NOTE: this will not run!

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Enable client to invoke the initialize and update operations separately [skip ci]

This commit adds support in the client for the separated initialize and
update operations on contracts. Specifically the python request and
response classes are now split into initialize request/response, update
request/response and no-change/error response.

NOTE: this will not run!

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add support for the split operations to CCF

This should complete the implementation of the
initialize and update operations.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Implements contract metadata and registers it with CCF

When a contract is initialized an encryption key pair and a signing key
pair are created and stored in privileged keys in state. A hash is
computed and passed through the contract initialization to CCF. CCF can
then provide an attestation to the veracity of the metadata hash for a
contract.  This ultimately enables contract to contract attestation.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Implement the initialize and update separation for Sawtooth

Fix signature computation for the initialize and
update operations for sawtooth. Currently the metadata
hash is not saved.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Change the construction of the code hash to enable code attestation

Move from a single hash of code to a merkle tree composed of the
code and name, and the nonce. This allows the nonce to be sufficient
(with the merkle root which is registered with CCF) to serve as
an attestation of the code.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Implment update contract & ledger commands for pdosh

Implement commands for interacting with the ledger (e.g.
to get the ledger verification key or ledger attested information
about the contracts or enclaves) and with a contract (e.g.
to get the contract id or provisioned enclaves). This enables
scripts to collect the information necessary to create a verifiable
attestation for a contract.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add a contract and scripts that test attestation capabililities.

Add test for contract metadata and state attestation.

This implements a demonstration of a (fully) functional commit/reveal
method for creating and storing a secret, then revealing it only if the
client can prove that the current state has been committed on the ledger

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* remove some LGTM warnings

Signed-off-by: Mic Bowman <mic.bowman@intel.com>