Add test for SGX report verification
Add a test for SGX report verification and add the test
to the build process. Data for the test is copied from
the common library.

Added an additional command to the pdo-shell to retrieve
state and configuration information. This makes it possible
to make the tests conditional on the ledger type. That is,
the attestation tests can only be performed if the ledger
is CCF.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>