Add to wawaka native functions to verify sgx attestation (#341)
* Fix const modifier for verify IAS report functions

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Fix the eservice db command and a few issues with enclave keys

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add Wawaka support for verifying the SGX attestation

Add native functions for verifying SGX attestation report.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* Add test for SGX report verification

Add a test for SGX report verification and add the test
to the build process. Data for the test is copied from
the common library.

Added an additional command to the pdo-shell to retrieve
state and configuration information. This makes it possible
to make the tests conditional on the ledger type. That is,
the attestation tests can only be performed if the ledger
is CCF.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* fix return type from verify_ias_report wawaka functions

Signed-off-by: Mic Bowman <mic.bowman@intel.com>

* small changes to address PR feedback

Signed-off-by: Mic Bowman <mic.bowman@intel.com>