Add wawaka native function to parse the sgx quote and report
Add a function, sgx_parse_report, to convert the binary sgx report
into a JSON structure that can be interpreted by the contract. Add
a new common function package for processing the sgx attestation.
Update the tests for the new values.

Also fix a bug in the error handling of pdo-shell.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>