Add library of common wawaka contract methods.
There are several methods that are common to many contracts. This
update provides a base implementation for those methods in order to
simplify contract development.

Add common contracts methods for attestations

Provides a wrapper for managing attested contract endpoints.
Contract object provides an attestation package that is verified
registering the encryption & verification keys for the contract.
Useful for building contracts that share secrets.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>