Push the block_store_manager use across client & server
Replace the file system block cache with lmdb-based block manager
Replace the block store manager from the sservice with the one common one
Push use of configuration variables into the appropriate locations

The storage service uses an LMDB-based block manager for keeping
data and metadata in an efficient store. This moves most of the
storage service functionality to the client enabling a more functional
local cache (that is, the metadata stored can be used to expire
old blocks similarly to the storage service.

The unification of a block manager layer also makes it easier
to sync blocks across different storage services.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>