Move eservice database creation & update to site.psh.
A small change to the the eservice command to load the
datatbase makes it possible to create and update the
eservice database directly from the site.psh script.
This should simplify creating the database. Note that
the site.psh is set up for the current standard test
environment. It should be customized with any site-specific
changes.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>