Overhaul CMake files in common to support client-only build
Signficant overhaul of the CMake files in common for the purpose of:
1) supporting a client-only build that does not require SGX
2) moving shared variables consistently in to CMakeVariables.txt
3) generally cleaning up (consistency, appropriate use) files

There are now three build configurations: BUILD_TRUSTED,
BUILD_UNTRUSTED and BUILD_CLIENT that are set when the
Makefiles are created. By default the configurations are
identical to previous builds with BUILD_TRUSTED and BUILD_UNTRUSTED
set by default.

Target libraries are designated by t-* for trusted, u-* for untrusted
and c-* for client. Libraries such as blockstore that are applicable
for all configurations do not have a specific designation.

Signed-off-by: Mic Bowman <mic.bowman@intel.com>